# API/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.DefinirHorariosRequest:
    properties:
      horarios:
        items:
          $ref: '#/definitions/controllers.UsuarioHorarioRequest'
        type: array
    required:
    - horarios
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        example: Erro na operação
        type: string
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        example: Operação realizada com sucesso
        type: string
    type: object
...
  controllers.UsuarioHorarioRequest:
    properties:
      ativo:
        example: true
        type: boolean
      dia_semana:
        example: 1
        maximum: 6
        minimum: 0
        type: integer
      horario_fim:
        example: "18:00"
        type: string
      horario_inicio:
        example: "08:00"
        type: string
    required:
    - dia_semana
    - horario_fim
    - horario_inicio
    type: object
  controllers.UsuarioHorarioResponse:
    properties:
      ativo:
        example: true
        type: boolean
      dia_semana:
        example: 1
        type: integer
      dia_semana_texto:
        example: Segunda-feira
        type: string
      horario_fim:
        example: "18:00"
        type: string
      horario_inicio:
        example: "08:00"
        type: string
      id:
        example: 1
        type: integer
      usuario_id:
        example: 1
        type: integer
    type: object
...
  controllers.UsuarioRequest:
    properties:
      email:
        example: joao@exemplo.com
        type: string
      nome:
        example: João Silva
        type: string
      senha:
        example: "123456"
        type: string
      tipo_usuario_id:
        example: 1
        type: integer
    required:
    - email
    - nome
    - senha
    - tipo_usuario_id
    type: object
  controllers.UsuarioResponse:
    properties:
      ativo:
        example: true
        type: boolean
      clinica_id:
        example: 1
        type: integer
      email:
        example: joao@exemplo.com
        type: string
      id:
        example: 1
        type: integer
      nome:
        example: João Silva
        type: string
      tipo_usuario_id:
        example: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API multiclínica com isolamento por tenant (clinic_id), RBAC por papel
    (ADM_GERAL, DONO, MEDICO, SECRETARIA), agendamento com validação de conflitos,
    prontuário com imutabilidade após 24h, auditoria LGPD e métricas de faturamento
    e no-show.
  title: API White Label — Clínicas
  version: "1.0"
paths:
...
  /usuarios:
    get:
      consumes:
      - application/json
      description: Lista todos os usuários da clínica
      parameters:
      - description: Filtrar apenas usuários ativos
        in: query
        name: ativos
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.UsuarioResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar usuários
      tags:
      - Usuários
    post:
      consumes:
      - application/json
      description: Cria um novo usuário na clínica
      parameters:
      - description: Dados do usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/controllers.UsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.UsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar usuário
      tags:
      - Usuários
  /usuarios/{id}:
    delete:
      consumes:
      - application/json
      description: Desativa um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Desativar usuário
      tags:
      - Usuários
    get:
      consumes:
      - application/json
      description: Busca um usuário específico pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar usuário por ID
      tags:
      - Usuários
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/controllers.UsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar usuário
      tags:
      - Usuários
  /usuarios/{id}/horarios:
    get:
      consumes:
      - application/json
      description: Retorna os horários de trabalho do usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.UsuarioHorarioResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar horários do usuário
      tags:
      - Usuários
    put:
      consumes:
      - application/json
      description: Define os horários de trabalho do usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Horários do usuário
        in: body
        name: horarios
        required: true
        schema:
          $ref: '#/definitions/controllers.DefinirHorariosRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Definir horários do usuário
      tags:
      - Usuários
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"